+++ scripts/add_satellite_dataset.py
@@ -1,12 +1,7 @@
-import json
 from argparse import ArgumentParser, Namespace
-from json import JSONDecodeError
 from pathlib import Path
 
-import source.satellite_dataset.config as sdcfg
-import source.utility as util
-import user.config as ucfg
-from source.satellite_dataset.typing import SatelliteDataset
+import source.satellite_dataset.add as sd_add
 
 
 def main() -> None:
@@ -29,7 +24,7 @@ def main() -> None:
     if dataset_name is None:
         dataset_name = input("Enter dataset name: ")
 
-    add_satellite_dataset(dataset_path, dataset_archive, dataset_name)
+    sd_add.add_satellite_dataset(dataset_path, dataset_archive, dataset_name)
 
 
 def parse_input_args() -> Namespace:
@@ -46,52 +41,6 @@ def parse_input_args() -> Namespace:
     arg_parser.add_argument("name", nargs="?", help="name to give to the dataset")
 
     return arg_parser.parse_args()
-
-
-def add_satellite_dataset(
-    dataset_path: Path, dataset_archive: str, dataset_name: str
-) -> None:
-    satellite_datasets_json_path = sdcfg.DATASETS_JSON_PATH
-    satellite_datasets_json_path.parent.mkdir(parents=True, exist_ok=True)
-
-    satellite_datasets: dict[str, SatelliteDataset] = {}
-
-    if satellite_datasets_json_path.is_file():
-        with open(satellite_datasets_json_path, "r") as json_file:
-            try:
-                satellite_datasets = json.load(json_file)
-            except JSONDecodeError:
-                print(
-                    f"Warning: JSON file '{satellite_datasets_json_path.as_posix()}' "
-                    "contains invalid syntax and could not be deserialized."
-                )
-
-                if not util.user_confirm(
-                    "Continue and overwrite existing file content?"
-                ):
-                    return
-
-        if dataset_name in satellite_datasets:
-            print(
-                f"Warning: '{satellite_datasets_json_path.as_posix()}' already "
-                f"contains a dataset with the name '{dataset_name}'."
-            )
-
-            if not util.user_confirm("Continue and overwrite existing dataset?"):
-                return
-
-    with open(satellite_datasets_json_path, "w") as json_file:
-        satellite_datasets[dataset_name] = {
-            "path": dataset_path.as_posix(),
-            "archive": dataset_archive,
-        }
-
-        json.dump(satellite_datasets, json_file, indent=ucfg.JSON_INDENT)
-
-        print(
-            f"Successfully added dataset '{dataset_name}' to "
-            f"{satellite_datasets_json_path.as_posix()}."
-        )
 
 
 if __name__ == "__main__":
